requirements.txt
colorama
requests
beautifulsoup4
fake_headers
websocket-client